requirements.txt
beautifulsoup4
arabic-reshaper
python-bidi
wordcloud
matplotlib

